+++ src/test/java/com/exercise/basic/array/Bai8Test.java
@@ -1,8 +1,8 @@
-package com.exercise.basic;
+package com.exercise.basic.array;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class Bai8Test {
